requirements.txt
flask==2.0.1
flask-cors==3.0.10
gTTS==2.3.1
gunicorn==20.1.0
selenium==4.1.0
webdriver-manager==3.8.0
python-dotenv==0.19.0 